Rename main file
"rect test" probably isn't the best name for the file that runs our
game.
